Forgot to half4 col as well

+++ Assets/Scenes/BenjaminSwee_Tutorial/Shaders/NewUnlitShader.shader
@@ -43,7 +43,7 @@ Shader "Custom/URP_BaseShader"
             half4 frag (v2f i) : SV_Target
             {
                 // sample the texture
-                fixed4 col = tex2D(_MainTex, i.uv);
+                half4 col = tex2D(_MainTex, i.uv);
                 InputData inputdata = (inputdata)0;
                 inputdata.positionWS = 0;
                 inputdata.normalWS = 0;
